pyproject.toml
[project]
name = "paperclassified"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hydra-core>=1.3.2",
    "kagglehub>=1.0.0",
    "kaleido>=1.2.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "onnx>=1.19.1",
    "onnxruntime>=1.24.4",
    "optuna>=4.8.0",
    "optuna-dashboard>=0.20.0",
    "pandas>=3.0.2",
    "pathlib>=1.0.1",
    "plotly>=6.6.0",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "torch>=2.11.0",
    "tqdm>=4.67.3",
    "transformers>=5.4.0",
    "wandb>=0.25.1",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "ipykernel>=7.2.0",
    "isort>=7.0.0",
    "pre-commit>=4.5.1",
    "ruff>=0.14.3",
]

[tool.black]
line-length = 88
target-version = ["py313"]
extend-exclude = '''
(
  ^outputs/
  | ^models/
  | ^data/
)
'''

[tool.ruff]
line-length = 88
target-version = "py313"
exclude = [
    ".git",
    ".venv",
    "data",
    "models",
    "outputs",
]

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.isort]
profile = "black"
line_length = 88
py_version = 313
skip = [".git", ".venv", "data", "models", "outputs"]
